fix: delete stuff.

+++ src/main/java/me/ryzeon/finanzas/service/impl/InvoiceServiceImpl.java
@@ -33,6 +33,9 @@ public class InvoiceServiceImpl implements InvoiceService {
 
     @Override
     public void deleteInvoice(Long id) {
+        if (!invoiceRepository.existsById(id)) {
+            throw new EntityNotFoundException("Invoice not found");
+        }
         invoiceRepository.deleteById(id);
     }
